makefriends/requirements.txt
django
djangorestframework
psycopg2
djangorestframework
djangorestframework_simplejwt
django-axes
django-two-factor-auth
django-otp-yubikey 
cryptography
django-phonenumbers
django-allauth
dj-rest-auth